refactor: rename TestCaseData files to TestData and add summaries where missing
