+++ HideAndSeekTestProject/TestGameController_LoadGame_TestData.cs
@@ -1,13 +1,16 @@
-﻿using Moq;
-using System.Collections; // From TestableIO
-using System.IO.Abstractions;
+﻿using System;
+using System.Collections;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
 
 namespace HideAndSeek
 {
     /// <summary>
-    /// TestCaseData for GameController tests for SaveGame method to save game to file
+    /// Test data for GameController tests for LoadGame method to load saved game from file
     /// </summary>
-    public static class TestGameController_SaveGame_TestCaseData
+    public static class TestGameController_LoadGame_TestData
     {
         /// <summary>
         /// Dictionary of Opponents and associated LocationWithHidingPlace names
@@ -250,7 +253,7 @@ namespace HideAndSeek
                     "}";
             }
         }
-         
+
         /// <summary>
         /// Enumerable of Location names in Locations property in default House for tests
         /// </summary>
@@ -325,7 +328,7 @@ namespace HideAndSeek
         {
             get
             {
-                return 
+                return
                 "{" +
                     "\"Name\":\"test house\"" + "," +
                     "\"HouseFileName\":\"TestHouse\"" + "," +
@@ -458,203 +461,319 @@ namespace HideAndSeek
             }
         }
 
-        public static IEnumerable TestCases_For_Test_GameController_SaveGame_AndCheckTextSavedToFile
+        public static IEnumerable TestCases_For_Test_GameController_LoadGame_WithNoFoundOpponents
         {
             get
             {
-                // Default House, no Opponents found
+                // Default House
                 yield return new TestCaseData(
-                        "DefaultHouse.json",
+                        "{" +
+                            "\"HouseFileName\":\"DefaultHouse\"" + "," +
+                            "\"PlayerLocation\":\"Entry\"" + "," +
+                            "\"MoveNumber\":1" + "," +
+                            SavedGame_Serialized_OpponentsAndHidingLocations + "," +
+                            "\"FoundOpponents\":[]" +
+                        "}",
+                        "my house",
+                        "DefaultHouse",
                         DefaultHouse_Serialized,
-                        (Mock<IFileSystem> mockHouseFileSystem) =>
-                        {
-                            // Create GameController with specified file system, hide all Opponents in specified locations, and return GameController
-                            return new GameController("DefaultHouse")
-                                       .RehideAllOpponents(SavedGame_OpponentsAndHidingPlaces.Values);
-                        }, 
-                        SavedGame_Serialized_NoFoundOpponents)
-                    .SetName("Test_GameController_SaveGame_AndCheckTextSavedToFile - default House - no opponents found")
-                    .SetCategory("GameController SaveGame Success");
+                        "Entry",
+                        DefaultHouse_Locations,
+                        DefaultHouse_LocationsWithoutHidingPlaces,
+                        DefaultHouse_LocationsWithHidingPlaces,
+                        "Entry",
+                        1,
+                        new List<string>() { "Joe", "Bob", "Ana", "Owen", "Jimmy" },
+                        new List<string>() { "Kitchen", "Pantry", "Bathroom", "Kitchen", "Pantry" })
+                    .SetName("Test_GameController_LoadGame_WithNoFoundOpponents - default House")
+                    .SetCategory("GameController LoadGame Success");
 
-                // Default House, 3 Opponents found
+                // Custom test House
                 yield return new TestCaseData(
-                        "DefaultHouse.json",
-                        DefaultHouse_Serialized,
-                        (Mock<IFileSystem> mockHouseFileSystem) => 
-                        {
-                            // Create GameController with specified file system and hide all Opponents in specified locations
-                            GameController gameController = new GameController("DefaultHouse")
-                                                                .RehideAllOpponents(SavedGame_OpponentsAndHidingPlaces.Values);
+                        "{" +
+                            "\"HouseFileName\":\"TestHouse\"" + "," +
+                            "\"PlayerLocation\":\"Landing\"" + "," +
+                            "\"MoveNumber\":1" + "," +
+                            "\"OpponentsAndHidingLocations\":" +
+                            "{" +
+                                "\"Joe\":\"Closet\"," +
+                                "\"Bob\":\"Yard\"," +
+                                "\"Ana\":\"Cellar\"," +
+                                "\"Owen\":\"Attic\"," +
+                                "\"Jimmy\":\"Yard\"" +
+                            "}" + "," +
+                            "\"FoundOpponents\":[]" +
+                        "}",
+                        "test house",
+                        "TestHouse",
+                        SerializedCustomTestHouse,
+                        "Landing",
+                        new List<string>() { "Landing", "Hallway", "Bedroom", "Closet", "Sensory Room", "Kitchen", "Cellar",
+                                             "Pantry", "Yard", "Bathroom", "Living Room", "Office", "Attic" },
+                        new List<string>() { "Landing", "Hallway" },
+                        new List<string>() { "Bedroom", "Closet", "Sensory Room", "Kitchen", "Cellar",
+                                             "Pantry", "Yard", "Bathroom", "Living Room", "Office", "Attic"},
+                        "Landing",
+                        1,
+                        new List<string>() { "Joe", "Bob", "Ana", "Owen", "Jimmy" },
+                        new List<string>() { "Closet", "Yard", "Cellar", "Attic", "Yard" })
+                    .SetName("Test_GameController_LoadGame_WithNoFoundOpponents - custom House")
+                    .SetCategory("GameController LoadGame CustomHouse Success");
+            }
+        }
 
-                            // Go to Kitchen and check to find 2 Opponents
-                            gameController.Move(Direction.East);
-                            gameController.Move(Direction.Northwest);;
-                            gameController.CheckCurrentLocation();
+        public static IEnumerable TestCases_For_Test_GameController_LoadGame_AndCheckErrorMessage_WhenSavedGameFileFormatIsInvalid
+        {
+            get
+            {
+                // RANDOM ISSUES
+                // No data in file
+                yield return new TestCaseData("The input does not contain any JSON tokens. Expected the input to start with a valid JSON token, when isFinalBlock is true. Path: $ | LineNumber: 0 | BytePositionInLine: 0.",
+                        "")
+                    .SetName("Test_GameController_LoadGame_AndCheckErrorMessage_WhenSavedGameFileFormatIsInvalid - no data in file");
 
-                            // Go to Bathroom and check to find 1 Opponent
-                            gameController.Move(Direction.Southeast);;
-                            gameController.Move(Direction.North);;
-                            gameController.CheckCurrentLocation();
+                // Only whitespace in file
+                yield return new TestCaseData("The input does not contain any JSON tokens. Expected the input to start with a valid JSON token, when isFinalBlock is true. Path: $ | LineNumber: 0 | BytePositionInLine: 2.",
+                        "  ")
+                    .SetName("Test_GameController_LoadGame_AndCheckErrorMessage_WhenSavedGameFileFormatIsInvalid - only whitespace in file");
 
-                            // Return GameController after 3 Opponents found
-                            return gameController;
-                        },
+                // Just characters in file (not JSON)
+                yield return new TestCaseData("'A' is an invalid start of a value. Path: $ | LineNumber: 0 | BytePositionInLine: 0.",
+                        "ABCDeaoueou[{}}({}")
+                    .SetName("Test_GameController_LoadGame_AndCheckErrorMessage_WhenSavedGameFileFormatIsInvalid - just characters in file");
+
+                // MISSING KEY/VALUE SET
+                // Missing House file name
+                yield return new TestCaseData("JSON deserialization for type 'HideAndSeek.SavedGame' was missing required properties, including the following: HouseFileName",
+                        "{" +
+                            SavedGame_Serialized_PlayerLocation_NoOpponentsGame + "," +
+                            SavedGame_Serialized_MoveNumber_NoFoundOpponents + "," +
+                            SavedGame_Serialized_OpponentsAndHidingLocations + "," +
+                            SavedGame_Serialized_FoundOpponents_NoFoundOpponents +
+                        "}")
+                    .SetName("Test_GameController_LoadGame_AndCheckErrorMessage_WhenSavedGameFileFormatIsInvalid - missing house file name");
+
+                // Missing player location
+                yield return new TestCaseData("JSON deserialization for type 'HideAndSeek.SavedGame' was missing required properties, including the following: PlayerLocation",
                         "{" +
                             SavedGame_Serialized_HouseFileName + "," +
-                            "\"PlayerLocation\":\"Bathroom\"" + "," +
-                            "\"MoveNumber\":7" + "," +
+                            SavedGame_Serialized_MoveNumber_NoFoundOpponents + "," +
                             SavedGame_Serialized_OpponentsAndHidingLocations + "," +
-                            "\"FoundOpponents\":[\"Joe\",\"Owen\",\"Ana\"]" +
+                            SavedGame_Serialized_FoundOpponents_NoFoundOpponents +
                         "}")
-                    .SetName("Test_GameController_SaveGame_AndCheckTextSavedToFile - default House - 3 opponents found")
-                    .SetCategory("GameController SaveGame Success");
-                
-                // Custom House with constructor, no Opponents found
-                yield return new TestCaseData(
-                        "TestHouse.json",
-                        SerializedCustomTestHouse,
-                        (Mock<IFileSystem> mockHouseFileSystem) =>
-                        {
-                            return new GameController("TestHouse") // Create GameController with specified file system and specific House
-                                       .RehideAllOpponents(new List<string>() { "Closet", "Yard", "Cellar", "Attic", "Yard" }); // and hide all Opponents in specified locations
-                        },
+                    .SetName("Test_GameController_LoadGame_AndCheckErrorMessage_WhenSavedGameFileFormatIsInvalid - missing player location");
+
+                // Missing move number
+                yield return new TestCaseData("JSON deserialization for type 'HideAndSeek.SavedGame' was missing required properties, including the following: MoveNumber",
                         "{" +
-                            "\"HouseFileName\":\"TestHouse\"" + "," +
-                            "\"PlayerLocation\":\"Landing\"" + "," +
-                            "\"MoveNumber\":1" + "," +
+                            SavedGame_Serialized_HouseFileName + "," +
+                            SavedGame_Serialized_PlayerLocation_NoOpponentsGame + "," +
+                            SavedGame_Serialized_OpponentsAndHidingLocations + "," +
+                            SavedGame_Serialized_FoundOpponents_NoFoundOpponents +
+                        "}")
+                    .SetName("Test_GameController_LoadGame_AndCheckErrorMessage_WhenSavedGameFileFormatIsInvalid - missing move number");
+
+                // Missing opponents and hiding locations
+                yield return new TestCaseData("JSON deserialization for type 'HideAndSeek.SavedGame' was missing required properties, including the following: OpponentsAndHidingLocations",
+                        "{" +
+                            SavedGame_Serialized_HouseFileName + "," +
+                            SavedGame_Serialized_PlayerLocation_NoOpponentsGame + "," +
+                            SavedGame_Serialized_MoveNumber_NoFoundOpponents + "," +
+                            SavedGame_Serialized_FoundOpponents_NoFoundOpponents +
+                        "}")
+                    .SetName("Test_GameController_LoadGame_AndCheckErrorMessage_WhenSavedGameFileFormatIsInvalid - missing opponents and hiding locations");
+
+                // Missing found opponents
+                yield return new TestCaseData("JSON deserialization for type 'HideAndSeek.SavedGame' was missing required properties, including the following: FoundOpponents",
+                        "{" +
+                            SavedGame_Serialized_HouseFileName + "," +
+                            SavedGame_Serialized_PlayerLocation_NoOpponentsGame + "," +
+                            SavedGame_Serialized_MoveNumber_NoFoundOpponents + "," +
+                            SavedGame_Serialized_OpponentsAndHidingLocations +
+                        "}")
+                    .SetName("Test_GameController_LoadGame_AndCheckErrorMessage_WhenSavedGameFileFormatIsInvalid - missing found opponents");
+            }
+        }
+
+        public static IEnumerable TestCases_For_Test_GameController_LoadGame_AndCheckErrorMessage_WhenSavedGameFileDataHasInvalidValue
+        {
+            get
+            {
+                // INVALID VALUE DATA
+                // Invalid House file name
+                yield return new TestCaseData("Cannot perform action because file name \"a8}{{ /@uaou12 \" is invalid (is empty or contains illegal characters, e.g. \\, /, or whitespace)",
+                        "{" +
+                            "\"HouseFileName\":\"a8}{{ /@uaou12 \"" + "," +
+                            SavedGame_Serialized_PlayerLocation_NoOpponentsGame + "," +
+                            SavedGame_Serialized_MoveNumber_NoFoundOpponents + "," +
+                            SavedGame_Serialized_OpponentsAndHidingLocations + "," +
+                            SavedGame_Serialized_FoundOpponents_NoFoundOpponents +
+                        "}")
+                    .SetName("Test_GameController_LoadGame_AndCheckErrorMessage_WhenSavedGameFileDataHasInvalidValue - invalid HouseFileName");
+
+                // Invalid player location
+                yield return new TestCaseData("invalid PlayerLocation",
+                        "{" +
+                            SavedGame_Serialized_HouseFileName + "," +
+                            "\"PlayerLocation\":\"Tree\"," +
+                            SavedGame_Serialized_MoveNumber_NoFoundOpponents + "," +
+                            SavedGame_Serialized_OpponentsAndHidingLocations + "," +
+                            SavedGame_Serialized_FoundOpponents_NoFoundOpponents +
+                        "}")
+                    .SetName("Test_GameController_LoadGame_AndCheckErrorMessage_WhenSavedGameFileDataHasInvalidValue - invalid PlayerLocation");
+
+                // Invalid (negative) move number
+                yield return new TestCaseData("invalid MoveNumber",
+                        "{" +
+                            SavedGame_Serialized_HouseFileName + "," +
+                            SavedGame_Serialized_PlayerLocation_NoOpponentsGame + "," +
+                            "\"MoveNumber\":-1" + "," +
+                            SavedGame_Serialized_OpponentsAndHidingLocations + "," +
+                            SavedGame_Serialized_FoundOpponents_NoFoundOpponents +
+                        "}")
+                    .SetName("Test_GameController_LoadGame_AndCheckErrorMessage_WhenSavedGameFileDataHasInvalidValue - invalid MoveNumber");
+
+                // No opponents
+                yield return new TestCaseData("no opponents",
+                        "{" +
+                            SavedGame_Serialized_HouseFileName + "," +
+                            SavedGame_Serialized_PlayerLocation_NoOpponentsGame + "," +
+                            SavedGame_Serialized_MoveNumber_NoFoundOpponents + "," +
+                            "\"OpponentsAndHidingLocations\":{}" + "," +
+                            SavedGame_Serialized_FoundOpponents_NoFoundOpponents +
+                        "}")
+                    .SetName("Test_GameController_LoadGame_AndCheckErrorMessage_WhenSavedGameFileDataHasInvalidValue - no opponents");
+
+                // Invalid hiding place for Joe (not yet found) because location does not exist
+                yield return new TestCaseData("invalid hiding location for opponent",
+                        "{" +
+                            SavedGame_Serialized_HouseFileName + "," +
+                            SavedGame_Serialized_PlayerLocation_NoOpponentsGame + "," +
+                            SavedGame_Serialized_MoveNumber_NoFoundOpponents + "," +
                             "\"OpponentsAndHidingLocations\":" +
                             "{" +
-                                "\"Joe\":\"Closet\"," +
-                                "\"Bob\":\"Yard\"," +
-                                "\"Ana\":\"Cellar\"," +
-                                "\"Owen\":\"Attic\"," +
-                                "\"Jimmy\":\"Yard\"" +
+                                "\"Joe\":\"Tree\"," +
+                                "\"Bob\":\"Pantry\"," +
+                                "\"Ana\":\"Bathroom\"," +
+                                "\"Owen\":\"Kitchen\"," +
+                                "\"Jimmy\":\"Pantry\"" +
                             "}" + "," +
-                            "\"FoundOpponents\":[]" +
+                            SavedGame_Serialized_FoundOpponents_NoFoundOpponents +
                         "}")
-                    .SetName("Test_GameController_SaveGame_AndCheckTextSavedToFile - custom House - constructor - no opponents found")
-                    .SetCategory("GameController SaveGame Success CustomHouse Constructor");
+                    .SetName("Test_GameController_LoadGame_AndCheckErrorMessage_WhenSavedGameFileDataHasInvalidValue - invalid hiding place for opponent - Location does not exist");
 
-                // Custom House with ReloadGame, no Opponents found
-                yield return new TestCaseData(
-                    "TestHouse.json",
-                        SerializedCustomTestHouse,
-                        (Mock<IFileSystem> mockHouseFileSystem) =>
-                        {
-                            // Add DefaultHouse file to House file system mock and set House file system
-                            mockHouseFileSystem = MockFileSystemHelper.SetMockOfFileSystem_ToReadAllText(
-                                                  mockHouseFileSystem, "DefaultHouse.json", DefaultHouse_Serialized);
-                            House.FileSystem = mockHouseFileSystem.Object;
-
-                            // Return GameController with restarted game and rehidden Opponents
-                            return new GameController("DefaultHouse") // Create GameController with specified file system
-                                       .RestartGame("TestHouse") // and restart game with specific House
-                                       .RehideAllOpponents(new List<string>() { "Closet", "Yard", "Cellar", "Attic", "Yard" }); // and hide all Opponents in specified locations
-                        },
+                // Invalid hiding place for Joe (not yet found) because hiding location is not of type LocationWithHidingPlace
+                yield return new TestCaseData("invalid hiding location for opponent",
                         "{" +
-                            "\"HouseFileName\":\"TestHouse\"" + "," +
-                            "\"PlayerLocation\":\"Landing\"" + "," +
-                            "\"MoveNumber\":1" + "," +
+                            SavedGame_Serialized_HouseFileName + "," +
+                            SavedGame_Serialized_PlayerLocation_NoOpponentsGame + "," +
+                            SavedGame_Serialized_MoveNumber_NoFoundOpponents + "," +
                             "\"OpponentsAndHidingLocations\":" +
                             "{" +
-                                "\"Joe\":\"Closet\"," +
-                                "\"Bob\":\"Yard\"," +
-                                "\"Ana\":\"Cellar\"," +
-                                "\"Owen\":\"Attic\"," +
-                                "\"Jimmy\":\"Yard\"" +
+                                "\"Joe\":\"Hallway\"," +
+                                "\"Bob\":\"Pantry\"," +
+                                "\"Ana\":\"Bathroom\"," +
+                                "\"Owen\":\"Kitchen\"," +
+                                "\"Jimmy\":\"Pantry\"" +
                             "}" + "," +
-                            "\"FoundOpponents\":[]" +
+                            SavedGame_Serialized_FoundOpponents_NoFoundOpponents +
                         "}")
-                    .SetName("Test_GameController_SaveGame_AndCheckTextSavedToFile - custom House - ReloadGame - no opponents found")
-                    .SetCategory("GameController SaveGame Success CustomHouse ReloadGame");
+                    .SetName("Test_GameController_LoadGame_AndCheckErrorMessage_WhenSavedGameFileDataHasInvalidValue - invalid hiding place for opponent - not LocationWithHidingPlace");
 
-                // Custom House with constructor, 3 Opponents found
-                yield return new TestCaseData(
-                        "TestHouse.json",
-                        SerializedCustomTestHouse,
-                        (Mock<IFileSystem> mockHouseFileSystem) =>
-                        {
-                            // Add DefaultHouse file to House file system mock and set House file system
-                            mockHouseFileSystem = MockFileSystemHelper.SetMockOfFileSystem_ToReadAllText(
-                                                  mockHouseFileSystem, "DefaultHouse.json", DefaultHouse_Serialized);
-                            House.FileSystem = mockHouseFileSystem.Object;
-
-                            // Initialize to GameController with restarted game and rehidden Opponents
-                            GameController gameController = new GameController("DefaultHouse") // Create GameController with specified file system and default House
-                                   .RestartGame("TestHouse") // and restart game with specific House
-                                   .RehideAllOpponents(new List<string>() { "Closet", "Yard", "Cellar", "Attic", "Yard" }); // and hide all Opponents in specified locations
-
-                            // Go to Cellar and find 1 Opponent there
-                            gameController.Move(Direction.North);;
-                            gameController.Move(Direction.East);
-                            gameController.Move(Direction.Down);;
-                            gameController.CheckCurrentLocation();
-
-                            // Go to Yard and find 2 Opponents there
-                            gameController.Move(Direction.Up);;
-                            gameController.Move(Direction.Out);;
-                            gameController.CheckCurrentLocation();
-
-                            // Return GameController
-                            return gameController;
-                        },
+                // Found opponent is not in all opponents list
+                yield return new TestCaseData("found opponent is not an opponent",
                         "{" +
-                            "\"HouseFileName\":\"TestHouse\"" + "," +
-                            "\"PlayerLocation\":\"Yard\"" + "," +
-                            "\"MoveNumber\":8" + "," +
-                            "\"OpponentsAndHidingLocations\":" +
-                            "{" +
-                                "\"Joe\":\"Closet\"," +
-                                "\"Bob\":\"Yard\"," +
-                                "\"Ana\":\"Cellar\"," +
-                                "\"Owen\":\"Attic\"," +
-                                "\"Jimmy\":\"Yard\"" +
-                            "}" + "," +
-                            "\"FoundOpponents\":[\"Ana\",\"Bob\",\"Jimmy\"]" +
+                            SavedGame_Serialized_HouseFileName + "," +
+                            SavedGame_Serialized_PlayerLocation_NoOpponentsGame + "," +
+                            SavedGame_Serialized_MoveNumber_NoFoundOpponents + "," +
+                            SavedGame_Serialized_OpponentsAndHidingLocations + "," +
+                            "\"FoundOpponents\":[\"Steve\"]" +
                         "}")
-                    .SetName("Test_GameController_SaveGame_AndCheckTextSavedToFile - custom House - constructor - 3 opponents found")
-                    .SetCategory("GameController SaveGame Success CustomHouse Constructor");
+                    .SetName("Test_GameController_LoadGame_AndCheckErrorMessage_WhenSavedGameFileDataHasInvalidValue - found opponent Steve is not opponent");
+            }
+        }
 
-                // Custom House with ReloadGame, 3 Opponents found
-                yield return new TestCaseData(
-                        "TestHouse.json",
-                        SerializedCustomTestHouse,
-                        (Mock<IFileSystem> mockHouseFileSystem) =>
+        public static IEnumerable TestCases_For_Test_GameController_LoadGame_WithFoundOpponents
+        {
+            get
+            {
+                // 1 found Opponent
+                yield return new TestCaseData("Bathroom", 4, new List<string>() { "Ana" },
+                        (IEnumerable<LocationWithHidingPlace> locationsWithHidingPlaces) =>
                         {
-                            // Initialize GameController
-                            GameController gameController = new GameController("TestHouse") // Create GameController with specified file system and specific House
-                                                                .RehideAllOpponents(new List<string>() { "Closet", "Yard", "Cellar", "Attic", "Yard" }); // and hide all Opponents in specified locations
+                            Assert.Multiple(() =>
+                            {
+                                // Assert that Opponents Joe and Owen are hiding in the Kitchen
+                                Assert.That(locationsWithHidingPlaces
+                                                .Where((l) => l.Name == "Kitchen").First()
+                                                .CheckHidingPlace()
+                                                .Select((o) => o.Name),
+                                            Is.EquivalentTo(new List<string>() { "Joe", "Owen" }),
+                                            "Joe and Owen are hidden in the Kitchen");
 
-                            // Go to Cellar and find 1 Opponent there
-                            gameController.Move(Direction.North);;
-                            gameController.Move(Direction.East);
-                            gameController.Move(Direction.Down);;
-                            gameController.CheckCurrentLocation();
+                                // Assert that Opponents Bob and Jimmy are hiding in the Pantry
+                                Assert.That(locationsWithHidingPlaces
+                                                .Where((l) => l.Name == "Pantry").First()
+                                                .CheckHidingPlace()
+                                                .Select((o) => o.Name),
+                                            Is.EquivalentTo(new List<string>() { "Bob", "Jimmy" }),
+                                            "Bob and Jimmy are hidden in the Pantry");
+                            });
+                        })
+                    .SetName("Test_GameController_LoadGame_WithFoundOpponents - 1 found opponent");
 
-                            // Go to Yard and find 2 Opponents there
-                            gameController.Move(Direction.Up);;
-                            gameController.Move(Direction.Out);;
-                            gameController.CheckCurrentLocation();
+                // 2 found Opponents
+                yield return new TestCaseData("Pantry", 10, new List<string>() { "Bob", "Jimmy" },
+                        (IEnumerable<LocationWithHidingPlace> locationsWithHidingPlaces) =>
+                        {
+                            Assert.Multiple(() =>
+                            {
+                                // Assert that Opponents Joe and Owen are hiding in the Kitchen
+                                Assert.That(locationsWithHidingPlaces
+                                                .Where((l) => l.Name == "Kitchen").First()
+                                                .CheckHidingPlace()
+                                                .Select((o) => o.Name),
+                                            Is.EquivalentTo(new List<string>() { "Joe", "Owen" }),
+                                            "Joe and Owen are hidden in the Kitchen");
 
-                            // Return GameController
-                            return gameController;
-                        },
-                        "{" +
-                            "\"HouseFileName\":\"TestHouse\"" + "," +
-                            "\"PlayerLocation\":\"Yard\"" + "," +
-                            "\"MoveNumber\":8" + "," +
-                            "\"OpponentsAndHidingLocations\":" +
-                            "{" +
-                                "\"Joe\":\"Closet\"," +
-                                "\"Bob\":\"Yard\"," +
-                                "\"Ana\":\"Cellar\"," +
-                                "\"Owen\":\"Attic\"," +
-                                "\"Jimmy\":\"Yard\"" +
-                            "}" + "," +
-                            "\"FoundOpponents\":[\"Ana\",\"Bob\",\"Jimmy\"]" +
-                        "}")
-                    .SetName("Test_GameController_SaveGame_AndCheckTextSavedToFile - custom House - ReloadGame - 3 opponents found")
-                    .SetCategory("GameController SaveGame Success CustomHouse ReloadGame");
+                                // Assert that Ana is hiding in the Bathroom
+                                Assert.That(locationsWithHidingPlaces
+                                                .Where((l) => l.Name == "Bathroom").First()
+                                                .CheckHidingPlace()
+                                                .Select((o) => o.Name),
+                                            Is.EquivalentTo(new List<string>() { "Ana" }),
+                                            "Ana is hidden in the Bathroom");
+                            });
+                        })
+                    .SetName("Test_GameController_LoadGame_WithFoundOpponents - 2 found opponents");
+
+                // 3 found Opponents
+                yield return new TestCaseData("Bathroom", 15, new List<string>() { "Joe", "Owen", "Ana" },
+                        (IEnumerable<LocationWithHidingPlace> locationsWithHidingPlaces) =>
+                        {
+                            // Assert that Opponents Joe and Owen are hiding in the Kitchen
+                            Assert.That(locationsWithHidingPlaces
+                                            .Where((l) => l.Name == "Pantry").First()
+                                            .CheckHidingPlace()
+                                            .Select((o) => o.Name),
+                                        Is.EquivalentTo(new List<string>() { "Bob", "Jimmy" }),
+                                        "Bob and Jimmy are hidden in the Kitchen");
+                        })
+                    .SetName("Test_GameController_LoadGame_WithFoundOpponents - 3 found opponents");
+
+                // 4 found Opponents
+                yield return new TestCaseData("Pantry", 20, new List<string>() { "Joe", "Owen", "Bob", "Jimmy" },
+                        (IEnumerable<LocationWithHidingPlace> locationsWithHidingPlaces) =>
+                        {
+                            // Assert that Ana is hiding in the Bathroom
+                            Assert.That(locationsWithHidingPlaces
+                                            .Where((l) => l.Name == "Bathroom").First()
+                                            .CheckHidingPlace()
+                                            .Select((o) => o.Name),
+                                        Is.EquivalentTo(new List<string>() { "Ana" }),
+                                        "Ana is hidden in the Bathroom");
+                        })
+                    .SetName("Test_GameController_LoadGame_WithFoundOpponents - 4 found opponents");
             }
         }
     }
