+++ HideAndSeekTestProject/TestSavedGame_SerializationAndDeserialization.cs
@@ -34,12 +34,12 @@ namespace HideAndSeek
         [Category("SavedGame Serialize Success")]
         public void Test_SavedGame_Serialize_NoFoundOpponents()
         {
-            SavedGame savedGame = new SavedGame(TestSavedGame_SerializationAndDeserialization_TestCaseData.GetDefaultHouse(), 
+            SavedGame savedGame = new SavedGame(TestSavedGame_SerializationAndDeserialization_TestData.GetDefaultHouse(), 
                                                 "DefaultHouse", "Entry", 1, 
-                                                TestSavedGame_SerializationAndDeserialization_TestCaseData.SavedGame_OpponentsAndHidingPlaces, 
+                                                TestSavedGame_SerializationAndDeserialization_TestData.SavedGame_OpponentsAndHidingPlaces, 
                                                 new List<string>());
             string serializedGame = JsonSerializer.Serialize(savedGame);
-            Assert.That(serializedGame, Is.EqualTo(TestSavedGame_SerializationAndDeserialization_TestCaseData.SavedGame_Serialized_NoFoundOpponents));
+            Assert.That(serializedGame, Is.EqualTo(TestSavedGame_SerializationAndDeserialization_TestData.SavedGame_Serialized_NoFoundOpponents));
         }
 
         // Tests all properties' getters
@@ -50,17 +50,17 @@ namespace HideAndSeek
             // Initialize variable to expected text for serialized SavedGame
             string expectedSavedGameText =
                 "{" +
-                    TestSavedGame_SerializationAndDeserialization_TestCaseData.SavedGame_Serialized_HouseFileName + "," +
+                    TestSavedGame_SerializationAndDeserialization_TestData.SavedGame_Serialized_HouseFileName + "," +
                     "\"PlayerLocation\":\"Bathroom\"" + "," +
                     "\"MoveNumber\":7" + "," +
-                    TestSavedGame_SerializationAndDeserialization_TestCaseData.SavedGame_Serialized_OpponentsAndHidingLocations + "," +
+                    TestSavedGame_SerializationAndDeserialization_TestData.SavedGame_Serialized_OpponentsAndHidingLocations + "," +
                     "\"FoundOpponents\":[\"Joe\",\"Owen\",\"Ana\"]" +
                 "}";
             
             // Create SavedGame object
-            SavedGame savedGame = new SavedGame(TestSavedGame_SerializationAndDeserialization_TestCaseData.GetDefaultHouse(), 
+            SavedGame savedGame = new SavedGame(TestSavedGame_SerializationAndDeserialization_TestData.GetDefaultHouse(), 
                                                 "DefaultHouse", "Bathroom", 7, 
-                                                TestSavedGame_SerializationAndDeserialization_TestCaseData.SavedGame_OpponentsAndHidingPlaces,
+                                                TestSavedGame_SerializationAndDeserialization_TestData.SavedGame_OpponentsAndHidingPlaces,
                                                 new List<string>() { "Joe", "Owen", "Ana" });
             
             // Serialize SavedGame object
@@ -77,16 +77,16 @@ namespace HideAndSeek
         {
             // Set mock file system to House property
             House.FileSystem = MockFileSystemHelper.GetMockedFileSystem_ToReadAllText("DefaultHouse.json", 
-                                    TestSavedGame_SerializationAndDeserialization_TestCaseData.DefaultHouse_Serialized);
+                                    TestSavedGame_SerializationAndDeserialization_TestData.DefaultHouse_Serialized);
 
             // Initialize variable to serialized SavedGame
             string savedGameFileText =
                 "{" +
-                    TestSavedGame_SerializationAndDeserialization_TestCaseData.SavedGame_Serialized_HouseFileName + "," +
-                    TestSavedGame_SerializationAndDeserialization_TestCaseData.SavedGame_Serialized_PlayerLocation_NoFoundOpponents + "," +
-                    TestSavedGame_SerializationAndDeserialization_TestCaseData.SavedGame_Serialized_MoveNumber_NoFoundOpponents + "," +
-                    TestSavedGame_SerializationAndDeserialization_TestCaseData.SavedGame_Serialized_OpponentsAndHidingLocations + "," +
-                    TestSavedGame_SerializationAndDeserialization_TestCaseData.SavedGame_Serialized_FoundOpponents_NoFoundOpponents +
+                    TestSavedGame_SerializationAndDeserialization_TestData.SavedGame_Serialized_HouseFileName + "," +
+                    TestSavedGame_SerializationAndDeserialization_TestData.SavedGame_Serialized_PlayerLocation_NoFoundOpponents + "," +
+                    TestSavedGame_SerializationAndDeserialization_TestData.SavedGame_Serialized_MoveNumber_NoFoundOpponents + "," +
+                    TestSavedGame_SerializationAndDeserialization_TestData.SavedGame_Serialized_OpponentsAndHidingLocations + "," +
+                    TestSavedGame_SerializationAndDeserialization_TestData.SavedGame_Serialized_FoundOpponents_NoFoundOpponents +
                 "}";
 
             // Attempt to deserialize text from file into SavedGame object
@@ -100,9 +100,9 @@ namespace HideAndSeek
                 Assert.That(savedGame.PlayerLocation, Is.EqualTo("Entry"), "player location");
                 Assert.That(savedGame.MoveNumber, Is.EqualTo(1), "move number");
                 Assert.That(savedGame.OpponentsAndHidingLocations.Select((kvp) => kvp.Key),
-                    Is.EquivalentTo(TestSavedGame_SerializationAndDeserialization_TestCaseData.SavedGame_OpponentsAndHidingPlaces.Select((kvp) => kvp.Key)), "names of opponents in opponents and hiding locations dictionary");
+                    Is.EquivalentTo(TestSavedGame_SerializationAndDeserialization_TestData.SavedGame_OpponentsAndHidingPlaces.Select((kvp) => kvp.Key)), "names of opponents in opponents and hiding locations dictionary");
                 Assert.That(savedGame.OpponentsAndHidingLocations.Select((kvp) => kvp.Value),
-                    Is.EquivalentTo(TestSavedGame_SerializationAndDeserialization_TestCaseData.SavedGame_OpponentsAndHidingPlaces.Select((kvp) => kvp.Value)), "names of hiding locations in opponents and hiding locations dictionary");
+                    Is.EquivalentTo(TestSavedGame_SerializationAndDeserialization_TestData.SavedGame_OpponentsAndHidingPlaces.Select((kvp) => kvp.Value)), "names of hiding locations in opponents and hiding locations dictionary");
                 Assert.That(savedGame.FoundOpponents, Is.Empty, "no found opponents");
 
                 // Assert that House properties are as expected
@@ -119,15 +119,15 @@ namespace HideAndSeek
         {
             // Set mock file system to House property
             House.FileSystem = MockFileSystemHelper.GetMockedFileSystem_ToReadAllText("DefaultHouse.json", 
-                                    TestSavedGame_SerializationAndDeserialization_TestCaseData.DefaultHouse_Serialized);
+                                    TestSavedGame_SerializationAndDeserialization_TestData.DefaultHouse_Serialized);
 
             // Initialize variable to serialized SavedGame
             string savedGameFileText =
                 "{" +
-                    TestSavedGame_SerializationAndDeserialization_TestCaseData.SavedGame_Serialized_HouseFileName + "," +
+                    TestSavedGame_SerializationAndDeserialization_TestData.SavedGame_Serialized_HouseFileName + "," +
                     "\"PlayerLocation\":\"Bathroom\"" + "," +
                     "\"MoveNumber\":7" + "," +
-                    TestSavedGame_SerializationAndDeserialization_TestCaseData.SavedGame_Serialized_OpponentsAndHidingLocations + "," +
+                    TestSavedGame_SerializationAndDeserialization_TestData.SavedGame_Serialized_OpponentsAndHidingLocations + "," +
                     "\"FoundOpponents\":[\"Joe\",\"Owen\",\"Ana\"]" +
                 "}";
 
@@ -142,9 +142,9 @@ namespace HideAndSeek
                 Assert.That(savedGame.PlayerLocation, Is.EqualTo("Bathroom"), "player location");
                 Assert.That(savedGame.MoveNumber, Is.EqualTo(7), "move number");
                 Assert.That(savedGame.OpponentsAndHidingLocations.Select((kvp) => kvp.Key),
-                    Is.EquivalentTo(TestSavedGame_SerializationAndDeserialization_TestCaseData.SavedGame_OpponentsAndHidingPlaces.Select((kvp) => kvp.Key)), "names of opponents in opponents and hiding locations dictionary");
+                    Is.EquivalentTo(TestSavedGame_SerializationAndDeserialization_TestData.SavedGame_OpponentsAndHidingPlaces.Select((kvp) => kvp.Key)), "names of opponents in opponents and hiding locations dictionary");
                 Assert.That(savedGame.OpponentsAndHidingLocations.Select((kvp) => kvp.Value),
-                    Is.EquivalentTo(TestSavedGame_SerializationAndDeserialization_TestCaseData.SavedGame_OpponentsAndHidingPlaces.Select((kvp) => kvp.Value)), "names of hiding locations in opponents and hiding locations dictionary");
+                    Is.EquivalentTo(TestSavedGame_SerializationAndDeserialization_TestData.SavedGame_OpponentsAndHidingPlaces.Select((kvp) => kvp.Value)), "names of hiding locations in opponents and hiding locations dictionary");
                 Assert.That(savedGame.FoundOpponents.Count(), Is.EqualTo(3), "3 found opponents");
                 Assert.That(savedGame.FoundOpponents, Is.EquivalentTo(new List<string>() { "Joe", "Owen", "Ana" }), "names of found opponents");
 
@@ -163,10 +163,10 @@ namespace HideAndSeek
             // Initialize variable to text representing serialized SavedGame
             string textInFile =
                 "{" +
-                    TestSavedGame_SerializationAndDeserialization_TestCaseData.SavedGame_Serialized_PlayerLocation_NoFoundOpponents + "," +
-                    TestSavedGame_SerializationAndDeserialization_TestCaseData.SavedGame_Serialized_MoveNumber_NoFoundOpponents + "," +
-                    TestSavedGame_SerializationAndDeserialization_TestCaseData.SavedGame_Serialized_OpponentsAndHidingLocations + "," +
-                    TestSavedGame_SerializationAndDeserialization_TestCaseData.SavedGame_Serialized_FoundOpponents_NoFoundOpponents +
+                    TestSavedGame_SerializationAndDeserialization_TestData.SavedGame_Serialized_PlayerLocation_NoFoundOpponents + "," +
+                    TestSavedGame_SerializationAndDeserialization_TestData.SavedGame_Serialized_MoveNumber_NoFoundOpponents + "," +
+                    TestSavedGame_SerializationAndDeserialization_TestData.SavedGame_Serialized_OpponentsAndHidingLocations + "," +
+                    TestSavedGame_SerializationAndDeserialization_TestData.SavedGame_Serialized_FoundOpponents_NoFoundOpponents +
                 "}";
 
             Assert.Multiple(() =>
@@ -182,7 +182,7 @@ namespace HideAndSeek
             });
         }
 
-        [TestCaseSource(typeof(TestSavedGame_SerializationAndDeserialization_TestCaseData), nameof(TestSavedGame_SerializationAndDeserialization_TestCaseData.TestCases_For_Test_SavedGame_Deserialize_AndCheckErrorMessage_ForJsonException_WhenNoJsonTokens))]
+        [TestCaseSource(typeof(TestSavedGame_SerializationAndDeserialization_TestData), nameof(TestSavedGame_SerializationAndDeserialization_TestData.TestCases_For_Test_SavedGame_Deserialize_AndCheckErrorMessage_ForJsonException_WhenNoJsonTokens))]
         [Category("SavedGame Deserialize Failure")]
         public void Test_SavedGame_Deserialize_AndCheckErrorMessage_ForJsonException_WhenNoJsonTokens(string expectedErrorMessage, string textInFile)
         {
@@ -199,13 +199,13 @@ namespace HideAndSeek
             });
         }
 
-        [TestCaseSource(typeof(TestSavedGame_SerializationAndDeserialization_TestCaseData), nameof(TestSavedGame_SerializationAndDeserialization_TestCaseData.TestCases_For_Test_SavedGame_Deserialize_AndCheckErrorMessage_ForJsonException_WhenPropertyIsMissing))]
+        [TestCaseSource(typeof(TestSavedGame_SerializationAndDeserialization_TestData), nameof(TestSavedGame_SerializationAndDeserialization_TestData.TestCases_For_Test_SavedGame_Deserialize_AndCheckErrorMessage_ForJsonException_WhenPropertyIsMissing))]
         [Category("SavedGame Deserialize Failure")]
         public void Test_SavedGame_Deserialize_AndCheckErrorMessage_ForJsonException_WhenPropertyIsMissing(string expectedErrorMessage, string textInFile)
         {
             // Set House file system to mock file system to return text in file
             House.FileSystem = MockFileSystemHelper.GetMockedFileSystem_ToReadAllText("DefaultHouse.json", 
-                                    TestSavedGame_SerializationAndDeserialization_TestCaseData.DefaultHouse_Serialized);
+                                    TestSavedGame_SerializationAndDeserialization_TestData.DefaultHouse_Serialized);
 
             Assert.Multiple(() =>
             {
@@ -220,13 +220,13 @@ namespace HideAndSeek
             });
         }
 
-        [TestCaseSource(typeof(TestSavedGame_SerializationAndDeserialization_TestCaseData), nameof(TestSavedGame_SerializationAndDeserialization_TestCaseData.TestCases_For_Test_SavedGame_Deserialize_AndCheckErrorMessage_ForInvalidDataException_WhenFileDataHasInvalidValue))]
+        [TestCaseSource(typeof(TestSavedGame_SerializationAndDeserialization_TestData), nameof(TestSavedGame_SerializationAndDeserialization_TestData.TestCases_For_Test_SavedGame_Deserialize_AndCheckErrorMessage_ForInvalidDataException_WhenFileDataHasInvalidValue))]
         [Category("SavedGame Deserialize Failure")]
         public void Test_SavedGame_Deserialize_AndCheckErrorMessage_ForInvalidDataException_WhenFileDataHasInvalidValue(string expectedErrorMessage, string textInFile)
         {
             // Set House file system to mock file system to return text in file
             House.FileSystem = MockFileSystemHelper.GetMockedFileSystem_ToReadAllText("DefaultHouse.json", 
-                                    TestSavedGame_SerializationAndDeserialization_TestCaseData.DefaultHouse_Serialized);
+                                    TestSavedGame_SerializationAndDeserialization_TestData.DefaultHouse_Serialized);
 
             Assert.Multiple(() =>
             {
@@ -241,7 +241,7 @@ namespace HideAndSeek
             });
         }
 
-        [TestCaseSource(typeof(TestSavedGame_SerializationAndDeserialization_TestCaseData), nameof(TestSavedGame_SerializationAndDeserialization_TestCaseData.TestCases_For_Test_SavedGame_Deserialize_AndCheckErrorMessage_ForJsonException_WhenHouseFileFormatIsInvalid))]
+        [TestCaseSource(typeof(TestSavedGame_SerializationAndDeserialization_TestData), nameof(TestSavedGame_SerializationAndDeserialization_TestData.TestCases_For_Test_SavedGame_Deserialize_AndCheckErrorMessage_ForJsonException_WhenHouseFileFormatIsInvalid))]
         [Category("SavedGame Deserialize House Failure")]
         public void Test_SavedGame_Deserialize_AndCheckErrorMessage_ForJsonException_WhenHouseFileFormatIsInvalid(string exceptionMessageEnding, string houseFileText)
         {
@@ -253,7 +253,7 @@ namespace HideAndSeek
                 // Assert that deserializing a SavedGame with a reference to a House with an invalid file format throws an exception
                 var exception = Assert.Throws<JsonException>(() =>
                 {
-                    JsonSerializer.Deserialize<SavedGame>(TestSavedGame_SerializationAndDeserialization_TestCaseData.SavedGame_Serialized_NoFoundOpponents);
+                    JsonSerializer.Deserialize<SavedGame>(TestSavedGame_SerializationAndDeserialization_TestData.SavedGame_Serialized_NoFoundOpponents);
                 });
 
                 // Assert that exception message is as expected
@@ -261,7 +261,7 @@ namespace HideAndSeek
             });
         }
 
-        [TestCaseSource(typeof(TestSavedGame_SerializationAndDeserialization_TestCaseData), nameof(TestSavedGame_SerializationAndDeserialization_TestCaseData.TestCases_For_Test_SavedGame_Deserialize_AndCheckErrorMessage_ForInvalidDataException_WhenHouseFileDataInvalidValue))]
+        [TestCaseSource(typeof(TestSavedGame_SerializationAndDeserialization_TestData), nameof(TestSavedGame_SerializationAndDeserialization_TestData.TestCases_For_Test_SavedGame_Deserialize_AndCheckErrorMessage_ForInvalidDataException_WhenHouseFileDataInvalidValue))]
         [Category("SavedGame Deserialize House Failure")]
         public void Test_SavedGame_Deserialize_AndCheckErrorMessage_ForInvalidDataException_WhenHouseFileDataInvalidValue(string exceptionMessageEnding, string houseFileText)
         {
@@ -273,7 +273,7 @@ namespace HideAndSeek
                 // Assert that deserializing a SavedGame with a reference to a House with invalid file data raises an exception
                 var exception = Assert.Throws<InvalidDataException>(() =>
                 {
-                    JsonSerializer.Deserialize<SavedGame>(TestSavedGame_SerializationAndDeserialization_TestCaseData.SavedGame_Serialized_NoFoundOpponents);
+                    JsonSerializer.Deserialize<SavedGame>(TestSavedGame_SerializationAndDeserialization_TestData.SavedGame_Serialized_NoFoundOpponents);
                 });
                 
                 // Assert that exception message is as expected
@@ -281,7 +281,7 @@ namespace HideAndSeek
             });
         }
 
-        [TestCaseSource(typeof(TestSavedGame_SerializationAndDeserialization_TestCaseData), nameof(TestSavedGame_SerializationAndDeserialization_TestCaseData.TestCases_For_Test_SavedGame_Deserialize_AndCheckErrorMessage_ForJsonException_WhenHouseFileDataHasInvalidDirection))]
+        [TestCaseSource(typeof(TestSavedGame_SerializationAndDeserialization_TestData), nameof(TestSavedGame_SerializationAndDeserialization_TestData.TestCases_For_Test_SavedGame_Deserialize_AndCheckErrorMessage_ForJsonException_WhenHouseFileDataHasInvalidDirection))]
         [Category("SavedGame Deserialize House Failure")]
         public void Test_SavedGame_Deserialize_AndCheckErrorMessage_ForJsonException_WhenHouseFileDataHasInvalidDirection(string houseFileText)
         {
@@ -293,7 +293,7 @@ namespace HideAndSeek
                 // Assert that deserializing a SavedGame with a reference to a House with an invalid Direction raises an exception
                 var exception = Assert.Throws<JsonException>(() =>
                 {
-                    JsonSerializer.Deserialize<SavedGame>(TestSavedGame_SerializationAndDeserialization_TestCaseData.SavedGame_Serialized_NoFoundOpponents);
+                    JsonSerializer.Deserialize<SavedGame>(TestSavedGame_SerializationAndDeserialization_TestData.SavedGame_Serialized_NoFoundOpponents);
                 });
 
                 // Assert that exception message is as expected
