+++ HideAndSeekTestProject/TestSavedGame_SerializationAndDeserialization_TestData.cs
@@ -1,7 +1,7 @@
-﻿using System;
+﻿using Newtonsoft.Json.Linq;
+using System;
 using System.Collections;
 using System.Collections.Generic;
-using System.IO.Abstractions;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -9,10 +9,163 @@ using System.Threading.Tasks;
 namespace HideAndSeek
 {
     /// <summary>
-    /// TestCaseData for some House tests for creating House objects with CreateHouse method
+    /// Test data for some SavedGame tests for serialization and deserialization
     /// </summary>
-    public static class TestHouse_TestCaseData
+    public static class TestSavedGame_SerializationAndDeserialization_TestData
     {
+        /// <summary>
+        /// Dictionary of Opponents and associated LocationWithHidingPlace names for SavedGame for tests
+        /// </summary>
+        public static Dictionary<string, string> SavedGame_OpponentsAndHidingPlaces
+        {
+            get
+            {
+                return new Dictionary<string, string>()
+                       {
+                           { "Joe", "Kitchen" },
+                           { "Bob", "Pantry" },
+                           { "Ana", "Bathroom" },
+                           { "Owen", "Kitchen" },
+                           { "Jimmy", "Pantry" }
+                       };
+            }
+        }
+
+        /// <summary>
+        /// Text for serialized test SavedGame with no found opponents
+        /// </summary>
+        public static string SavedGame_Serialized_NoFoundOpponents
+        {
+            get
+            {
+                return
+                    "{" +
+                        SavedGame_Serialized_HouseFileName + "," +
+                        SavedGame_Serialized_PlayerLocation_NoFoundOpponents + "," +
+                        SavedGame_Serialized_MoveNumber_NoFoundOpponents + "," +
+                        SavedGame_Serialized_OpponentsAndHidingLocations + "," +
+                        SavedGame_Serialized_FoundOpponents_NoFoundOpponents +
+                    "}";
+            }
+        }
+
+        /// <summary>
+        /// Text for serialized HouseFileName property of test SavedGame
+        /// </summary>
+        public static string SavedGame_Serialized_HouseFileName
+        {
+            get
+            {
+                return "\"HouseFileName\":\"DefaultHouse\"";
+            }
+        }
+
+        /// <summary>
+        /// Text for serialized PlayerLocation property of test SavedGame with no found opponents
+        /// </summary>
+        public static string SavedGame_Serialized_PlayerLocation_NoFoundOpponents
+        {
+            get
+            {
+                return "\"PlayerLocation\":\"Entry\"";
+            }
+        }
+
+        /// <summary>
+        /// Text for serialized MoveNumber property of test SavedGame with no found opponents
+        /// </summary>
+        public static string SavedGame_Serialized_MoveNumber_NoFoundOpponents
+        {
+            get
+            {
+                return "\"MoveNumber\":1";
+            }
+        }
+
+        /// <summary>
+        /// Text for serialized OpponentsAndHidingLocations property of test SavedGame
+        /// </summary>
+        public static string SavedGame_Serialized_OpponentsAndHidingLocations
+        {
+            get
+            {
+                return
+                    "\"OpponentsAndHidingLocations\":" +
+                    "{" +
+                        "\"Joe\":\"Kitchen\"," +
+                        "\"Bob\":\"Pantry\"," +
+                        "\"Ana\":\"Bathroom\"," +
+                        "\"Owen\":\"Kitchen\"," +
+                        "\"Jimmy\":\"Pantry\"" +
+                    "}";
+            }
+        }
+
+        /// <summary>
+        /// Text for serialized FoundOpponents property of test SavedGame with no found opponents
+        /// </summary>
+        public static string SavedGame_Serialized_FoundOpponents_NoFoundOpponents
+        {
+            get
+            {
+                return "\"FoundOpponents\":[]";
+            }
+        }
+
+        /// <summary>
+        /// Get new House object for testing purposes
+        /// </summary>
+        /// <returns>House object for testing purposes</returns>
+        public static House GetDefaultHouse()
+        {
+            // Create Entry and connect to new locations: Garage, Hallway
+            Location entry = new Location("Entry");
+            LocationWithHidingPlace garage = entry.AddExit(Direction.Out, "Garage", "behind the car");
+            Location hallway = entry.AddExit(Direction.East, "Hallway");
+
+            // Connect Hallway to new locations: Kitchen, Bathroom, Living Room, Landing
+            LocationWithHidingPlace kitchen = hallway.AddExit(Direction.Northwest, "Kitchen", "next to the stove");
+            LocationWithHidingPlace bathroom = hallway.AddExit(Direction.North, "Bathroom", "behind the door");
+            LocationWithHidingPlace livingRoom = hallway.AddExit(Direction.South, "Living Room", "behind the sofa");
+            Location landing = hallway.AddExit(Direction.Up, "Landing");
+
+            // Connect Landing to new locations: Attic, Kids Room, Master Bedroom, Nursery, Pantry, Second Bathroom
+            LocationWithHidingPlace attic = landing.AddExit(Direction.Up, "Attic", "in a trunk");
+            LocationWithHidingPlace kidsRoom = landing.AddExit(Direction.Southeast, "Kids Room", "in the bunk beds");
+            LocationWithHidingPlace masterBedroom = landing.AddExit(Direction.Northwest, "Master Bedroom", "under the bed");
+            LocationWithHidingPlace nursery = landing.AddExit(Direction.Southwest, "Nursery", "behind the changing table");
+            LocationWithHidingPlace pantry = landing.AddExit(Direction.South, "Pantry", "inside a cabinet");
+            LocationWithHidingPlace secondBathroom = landing.AddExit(Direction.West, "Second Bathroom", "in the shower");
+
+            // Connect Master Bedroom to new location: Master Bath
+            LocationWithHidingPlace masterBath = masterBedroom.AddExit(Direction.East, "Master Bath", "in the tub");
+
+            // Create list of Location objects (no hiding places)
+            IEnumerable<Location> locationsWithoutHidingPlaces = new List<Location>()
+            {
+                hallway, landing, entry
+            };
+
+            // Create list of LocationWithHidingPlace objects
+            IEnumerable<LocationWithHidingPlace> locationsWithHidingPlaces = new List<LocationWithHidingPlace>()
+            {
+                attic,
+                bathroom,
+                kidsRoom,
+                masterBedroom,
+                nursery,
+                pantry,
+                secondBathroom,
+                kitchen,
+                masterBath,
+                garage,
+                livingRoom
+            };
+
+            // Create and return new House
+            return new House("my house", "DefaultHouse", "Entry", locationsWithoutHidingPlaces, locationsWithHidingPlaces);
+        }
+
         /// <summary>
         /// Text representing default House for tests serialized
         /// </summary>
@@ -334,128 +487,181 @@ namespace HideAndSeek
             }
         }
 
-        /// <summary>
-        /// Enumerable of Location names in Locations property in default House for tests
-        /// </summary>
-        public static IEnumerable<string> DefaultHouse_Locations
+        public static IEnumerable TestCases_For_Test_SavedGame_Deserialize_AndCheckErrorMessage_ForJsonException_WhenNoJsonTokens
         {
             get
             {
-                return new List<string>()
-                    {
-                        "Attic",
-                        "Hallway",
-                        "Bathroom",
-                        "Kids Room",
-                        "Master Bedroom",
-                        "Nursery",
-                        "Pantry",
-                        "Second Bathroom",
-                        "Kitchen",
-                        "Master Bath",
-                        "Garage",
-                        "Landing",
-                        "Living Room",
-                        "Entry"
-                    };
+                // No data in file
+                yield return new TestCaseData("The input does not contain any JSON tokens. " +
+                                              "Expected the input to start with a valid JSON token, when isFinalBlock is true. " +
+                                              "Path: $ | LineNumber: 0 | BytePositionInLine: 0.",
+                        "")
+                    .SetName("Test_SavedGame_Deserialize_AndCheckErrorMessage_ForJsonException_WhenNoJsonTokens - no data in file");
+
+                // Only whitespace in file
+                yield return new TestCaseData("The input does not contain any JSON tokens. " +
+                                              "Expected the input to start with a valid JSON token, when isFinalBlock is true. " +
+                                              "Path: $ | LineNumber: 0 | BytePositionInLine: 2.",
+                        "  ")
+                    .SetName("Test_SavedGame_Deserialize_AndCheckErrorMessage_ForJsonException_WhenNoJsonTokens - only whitespace in file");
+
+                // Just characters in file (not JSON)
+                yield return new TestCaseData("'A' is an invalid start of a value. Path: $ | LineNumber: 0 | BytePositionInLine: 0.",
+                        "ABCDeaoueou[{}}({}")
+                    .SetName("Test_SavedGame_Deserialize_AndCheckErrorMessage_ForJsonException_WhenNoJsonTokens - just characters in file");
             }
         }
 
-        /// <summary>
-        /// Enumerable of Location names in LocationsWithoutHidingPlaces property in default House for tests
-        /// </summary>
-        public static IEnumerable<string> DefaultHouse_LocationsWithoutHidingPlaces
+        public static IEnumerable TestCases_For_Test_SavedGame_Deserialize_AndCheckErrorMessage_ForJsonException_WhenPropertyIsMissing
         {
             get
             {
-                return new List<string>()
-                        {
-                            "Hallway",
-                            "Landing",
-                            "Entry"
-                        };
+                // Missing player location
+                yield return new TestCaseData("JSON deserialization for type 'HideAndSeek.SavedGame' was missing required properties, " +
+                                              "including the following: PlayerLocation",
+                        "{" +
+                            SavedGame_Serialized_HouseFileName + "," +
+                            SavedGame_Serialized_MoveNumber_NoFoundOpponents + "," +
+                            SavedGame_Serialized_OpponentsAndHidingLocations + "," +
+                            SavedGame_Serialized_FoundOpponents_NoFoundOpponents +
+                        "}")
+                    .SetName("Test_SavedGame_Deserialize_AndCheckErrorMessage_ForJsonException_WhenPropertyIsMissing - missing player location");
+
+                // Missing move number
+                yield return new TestCaseData("JSON deserialization for type 'HideAndSeek.SavedGame' was missing required properties, " +
+                                              "including the following: MoveNumber",
+                        "{" +
+                            SavedGame_Serialized_HouseFileName + "," +
+                            SavedGame_Serialized_PlayerLocation_NoFoundOpponents + "," +
+                            SavedGame_Serialized_OpponentsAndHidingLocations + "," +
+                            SavedGame_Serialized_FoundOpponents_NoFoundOpponents +
+                        "}")
+                    .SetName("Test_SavedGame_Deserialize_AndCheckErrorMessage_ForJsonException_WhenPropertyIsMissing - missing move number");
+
+                // Missing opponents and hiding locations
+                yield return new TestCaseData("JSON deserialization for type 'HideAndSeek.SavedGame' was missing required properties, " +
+                                              "including the following: OpponentsAndHidingLocations",
+                        "{" +
+                            SavedGame_Serialized_HouseFileName + "," +
+                            SavedGame_Serialized_PlayerLocation_NoFoundOpponents + "," +
+                            SavedGame_Serialized_MoveNumber_NoFoundOpponents + "," +
+                            SavedGame_Serialized_FoundOpponents_NoFoundOpponents +
+                        "}")
+                    .SetName("Test_SavedGame_Deserialize_AndCheckErrorMessage_ForJsonException_WhenPropertyIsMissing - missing opponents and hiding locations");
+
+                // Missing found opponents
+                yield return new TestCaseData("JSON deserialization for type 'HideAndSeek.SavedGame' was missing required properties, " +
+                                              "including the following: FoundOpponents",
+                        "{" +
+                            SavedGame_Serialized_HouseFileName + "," +
+                            SavedGame_Serialized_PlayerLocation_NoFoundOpponents + "," +
+                            SavedGame_Serialized_MoveNumber_NoFoundOpponents + "," +
+                            SavedGame_Serialized_OpponentsAndHidingLocations +
+                        "}")
+                    .SetName("Test_SavedGame_Deserialize_AndCheckErrorMessage_ForJsonException_WhenPropertyIsMissing - missing found opponents");
+
             }
         }
 
-        /// <summary>
-        /// Enumerable of Location names in LocationsWithHidingPlaces property in default House for tests
-        /// </summary>
-        public static IEnumerable<string> DefaultHouse_LocationsWithHidingPlaces
+        public static IEnumerable TestCases_For_Test_SavedGame_Deserialize_AndCheckErrorMessage_ForInvalidDataException_WhenFileDataHasInvalidValue
         {
             get
             {
-                return new List<string>()
-                    {
-                        "Attic",
-                        "Bathroom",
-                        "Kids Room",
-                        "Master Bedroom",
-                        "Nursery",
-                        "Pantry",
-                        "Second Bathroom",
-                        "Kitchen",
-                        "Master Bath",
-                        "Garage",
-                        "Living Room",
-                    };
+                // Invalid House file name
+                yield return new TestCaseData("Cannot perform action because file name \"a8}{{ /@uaou12 \" is invalid " +
+                                              "(is empty or contains illegal characters, e.g. \\, /, or whitespace)",
+                        "{" +
+                            "\"HouseFileName\":\"a8}{{ /@uaou12 \"" + "," +
+                            SavedGame_Serialized_PlayerLocation_NoFoundOpponents + "," +
+                            SavedGame_Serialized_MoveNumber_NoFoundOpponents + "," +
+                            SavedGame_Serialized_OpponentsAndHidingLocations + "," +
+                            SavedGame_Serialized_FoundOpponents_NoFoundOpponents +
+                        "}")
+                    .SetName("Test_SavedGame_Deserialize_AndCheckErrorMessage_ForInvalidDataException_WhenFileDataHasInvalidValue - invalid HouseFileName");
+
+                // Invalid player location
+                yield return new TestCaseData("invalid PlayerLocation",
+                        "{" +
+                            SavedGame_Serialized_HouseFileName + "," +
+                            "\"PlayerLocation\":\"Tree\"," +
+                            SavedGame_Serialized_MoveNumber_NoFoundOpponents + "," +
+                            SavedGame_Serialized_OpponentsAndHidingLocations + "," +
+                            SavedGame_Serialized_FoundOpponents_NoFoundOpponents +
+                        "}")
+                    .SetName("Test_SavedGame_Deserialize_AndCheckErrorMessage_ForInvalidDataException_WhenFileDataHasInvalidValue - invalid PlayerLocation");
+
+                // Invalid (negative) move number
+                yield return new TestCaseData("invalid MoveNumber",
+                        "{" +
+                            SavedGame_Serialized_HouseFileName + "," +
+                            SavedGame_Serialized_PlayerLocation_NoFoundOpponents + "," +
+                            "\"MoveNumber\":-1" + "," +
+                            SavedGame_Serialized_OpponentsAndHidingLocations + "," +
+                            SavedGame_Serialized_FoundOpponents_NoFoundOpponents +
+                        "}")
+                    .SetName("Test_SavedGame_Deserialize_AndCheckErrorMessage_ForInvalidDataException_WhenFileDataHasInvalidValue - invalid MoveNumber");
+
+                // No opponents
+                yield return new TestCaseData("no opponents",
+                        "{" +
+                            SavedGame_Serialized_HouseFileName + "," +
+                            SavedGame_Serialized_PlayerLocation_NoFoundOpponents + "," +
+                            SavedGame_Serialized_MoveNumber_NoFoundOpponents + "," +
+                            "\"OpponentsAndHidingLocations\":{}" + "," +
+                            SavedGame_Serialized_FoundOpponents_NoFoundOpponents +
+                        "}")
+                    .SetName("Test_SavedGame_Deserialize_AndCheckErrorMessage_ForInvalidDataException_WhenFileDataHasInvalidValue - no opponents");
+
+                // Invalid hiding place for Joe (not yet found) because location does not exist
+                yield return new TestCaseData("invalid hiding location for opponent",
+                        "{" +
+                            SavedGame_Serialized_HouseFileName + "," +
+                            SavedGame_Serialized_PlayerLocation_NoFoundOpponents + "," +
+                            SavedGame_Serialized_MoveNumber_NoFoundOpponents + "," +
+                            "\"OpponentsAndHidingLocations\":" +
+                            "{" +
+                                "\"Joe\":\"Tree\"," +
+                                "\"Bob\":\"Pantry\"," +
+                                "\"Ana\":\"Bathroom\"," +
+                                "\"Owen\":\"Kitchen\"," +
+                                "\"Jimmy\":\"Pantry\"" +
+                            "}" + "," +
+                            SavedGame_Serialized_FoundOpponents_NoFoundOpponents +
+                        "}")
+                    .SetName("Test_SavedGame_Deserialize_AndCheckErrorMessage_ForInvalidDataException_WhenFileDataHasInvalidValue - invalid hiding place for opponent - Location does not exist");
+
+                // Invalid hiding place for Joe (not yet found) because hiding location is not of type LocationWithHidingPlace
+                yield return new TestCaseData("invalid hiding location for opponent",
+                        "{" +
+                            SavedGame_Serialized_HouseFileName + "," +
+                            SavedGame_Serialized_PlayerLocation_NoFoundOpponents + "," +
+                            SavedGame_Serialized_MoveNumber_NoFoundOpponents + "," +
+                            "\"OpponentsAndHidingLocations\":" +
+                            "{" +
+                                "\"Joe\":\"Hallway\"," +
+                                "\"Bob\":\"Pantry\"," +
+                                "\"Ana\":\"Bathroom\"," +
+                                "\"Owen\":\"Kitchen\"," +
+                                "\"Jimmy\":\"Pantry\"" +
+                            "}" + "," +
+                            SavedGame_Serialized_FoundOpponents_NoFoundOpponents +
+                        "}")
+                    .SetName("Test_SavedGame_Deserialize_AndCheckErrorMessage_ForInvalidDataException_WhenFileDataHasInvalidValue - invalid hiding place for opponent - not LocationWithHidingPlace");
+
+                // Found opponent is not in all opponents list
+                yield return new TestCaseData("found opponent is not an opponent",
+                        "{" +
+                            SavedGame_Serialized_HouseFileName + "," +
+                            SavedGame_Serialized_PlayerLocation_NoFoundOpponents + "," +
+                            SavedGame_Serialized_MoveNumber_NoFoundOpponents + "," +
+                            SavedGame_Serialized_OpponentsAndHidingLocations + "," +
+                            "\"FoundOpponents\":[\"Steve\"]" +
+                        "}")
+                    .SetName("Test_SavedGame_Deserialize_AndCheckErrorMessage_ForInvalidDataException_WhenFileDataHasInvalidValue - found opponent Steve is not opponent");
             }
         }
 
-        /// <summary>
-        /// Get new House object for testing purposes
-        /// </summary>
-        /// <returns>House object for testing purposes</returns>
-        public static House GetDefaultHouse()
-        {
-            // Create Entry and connect to new locations: Garage, Hallway
-            Location entry = new Location("Entry");
-            LocationWithHidingPlace garage = entry.AddExit(Direction.Out, "Garage", "behind the car");
-            Location hallway = entry.AddExit(Direction.East, "Hallway");
-
-            // Connect Hallway to new locations: Kitchen, Bathroom, Living Room, Landing
-            LocationWithHidingPlace kitchen = hallway.AddExit(Direction.Northwest, "Kitchen", "next to the stove");
-            LocationWithHidingPlace bathroom = hallway.AddExit(Direction.North, "Bathroom", "behind the door");
-            LocationWithHidingPlace livingRoom = hallway.AddExit(Direction.South, "Living Room", "behind the sofa");
-            Location landing = hallway.AddExit(Direction.Up, "Landing");
-
-            // Connect Landing to new locations: Attic, Kids Room, Master Bedroom, Nursery, Pantry, Second Bathroom
-            LocationWithHidingPlace attic = landing.AddExit(Direction.Up, "Attic", "in a trunk");
-            LocationWithHidingPlace kidsRoom = landing.AddExit(Direction.Southeast, "Kids Room", "in the bunk beds");
-            LocationWithHidingPlace masterBedroom = landing.AddExit(Direction.Northwest, "Master Bedroom", "under the bed");
-            LocationWithHidingPlace nursery = landing.AddExit(Direction.Southwest, "Nursery", "behind the changing table");
-            LocationWithHidingPlace pantry = landing.AddExit(Direction.South, "Pantry", "inside a cabinet");
-            LocationWithHidingPlace secondBathroom = landing.AddExit(Direction.West, "Second Bathroom", "in the shower");
-
-            // Connect Master Bedroom to new location: Master Bath
-            LocationWithHidingPlace masterBath = masterBedroom.AddExit(Direction.East, "Master Bath", "in the tub");
-
-            // Create list of Location objects (no hiding places)
-            IEnumerable<Location> locationsWithoutHidingPlaces = new List<Location>()
-            {
-                hallway, landing, entry
-            };
-
-            // Create list of LocationWithHidingPlace objects
-            IEnumerable<LocationWithHidingPlace> locationsWithHidingPlaces = new List<LocationWithHidingPlace>()
-            {
-                attic,
-                bathroom,
-                kidsRoom,
-                masterBedroom,
-                nursery,
-                pantry,
-                secondBathroom,
-                kitchen,
-                masterBath,
-                garage,
-                livingRoom
-            };
-
-            // Create and return new House
-            return new House("my house", "DefaultHouse", "Entry", locationsWithoutHidingPlaces, locationsWithHidingPlaces);
-        }
-
-        public static IEnumerable TestCases_For_Test_House_CreateHouse_AndCheckErrorMessage_ForJsonException_WhenFileFormatIsInvalid
+        public static IEnumerable TestCases_For_Test_SavedGame_Deserialize_AndCheckErrorMessage_ForJsonException_WhenHouseFileFormatIsInvalid
         {
             get
             {
@@ -465,79 +671,86 @@ namespace HideAndSeek
                                               "Expected the input to start with a valid JSON token, when isFinalBlock is true. " +
                                               "Path: $ | LineNumber: 0 | BytePositionInLine: 0.",
                         "")
-                    .SetName("Test_House_CreateHouse_AndCheckErrorMessage_ForJsonException_WhenFileFormatIsInvalid - empty file");
+                    .SetName("Test_SavedGame_Deserialize_AndCheckErrorMessage_ForJsonException_WhenHouseFileFormatIsInvalid - empty file");
 
                 // File with only whitespace
                 yield return new TestCaseData("The input does not contain any JSON tokens. " +
                                               "Expected the input to start with a valid JSON token, when isFinalBlock is true. " +
                                               "Path: $ | LineNumber: 0 | BytePositionInLine: 1.",
                         " ")
-                    .SetName("Test_House_CreateHouse_AndCheckErrorMessage_ForJsonException_WhenFileFormatIsInvalid - only whitespace");
+                    .SetName("Test_SavedGame_Deserialize_AndCheckErrorMessage_ForJsonException_WhenHouseFileFormatIsInvalid - only whitespace");
 
                 // File with random letters and characters
                 yield return new TestCaseData("'A' is an invalid start of a value. Path: $ | LineNumber: 0 | BytePositionInLine: 0.",
                         "ABCDeaoueou[{}}({}")
-                    .SetName("Test_House_CreateHouse_AndCheckErrorMessage_ForJsonException_WhenFileFormatIsInvalid - random characters");
+                    .SetName("Test_SavedGame_Deserialize_AndCheckErrorMessage_ForJsonException_WhenHouseFileFormatIsInvalid - random characters");
 
                 // MISSING PROPERTIES
                 // File missing Name
-                yield return new TestCaseData("JSON deserialization for type 'HideAndSeek.House' was missing required properties, including the following: Name",
+                yield return new TestCaseData("JSON deserialization for type 'HideAndSeek.House' was missing required properties, " +
+                                              "including the following: Name",
                         "{" +
                             DefaultHouse_Serialized_HouseFileName + "," +
                             DefaultHouse_Serialized_PlayerStartingPoint + "," +
                             DefaultHouse_Serialized_LocationsWithoutHidingPlaces + "," +
                             DefaultHouse_Serialized_LocationsWithHidingPlaces +
                         "}")
-                    .SetName("Test_House_CreateHouse_AndCheckErrorMessage_ForJsonException_WhenFileFormatIsInvalid - no Name");
+                    .SetName("Test_SavedGame_Deserialize_AndCheckErrorMessage_ForJsonException_WhenHouseFileFormatIsInvalid - no Name");
 
                 // File missing HouseFileName
-                yield return new TestCaseData("JSON deserialization for type 'HideAndSeek.House' was missing required properties, including the following: HouseFileName",
+                yield return new TestCaseData("JSON deserialization for type 'HideAndSeek.House' was missing required properties, " +
+                                              "including the following: HouseFileName",
                         "{" +
                            DefaultHouse_Serialized_Name + "," +
                            DefaultHouse_Serialized_PlayerStartingPoint + "," +
                            DefaultHouse_Serialized_LocationsWithoutHidingPlaces + "," +
                            DefaultHouse_Serialized_LocationsWithHidingPlaces +
                        "}")
-                   .SetName("Test_House_CreateHouse_AndCheckErrorMessage_ForJsonException_WhenFileFormatIsInvalid - no HouseFileName");
+                   .SetName("Test_SavedGame_Deserialize_AndCheckErrorMessage_ForJsonException_WhenHouseFileFormatIsInvalid - no HouseFileName");
 
                 // File missing PlayerStartingPoint
-                yield return new TestCaseData("JSON deserialization for type 'HideAndSeek.House' was missing required properties, including the following: PlayerStartingPoint",
+                yield return new TestCaseData("JSON deserialization for type 'HideAndSeek.House' was missing required properties, " +
+                                              "including the following: PlayerStartingPoint",
                         "{" +
                            DefaultHouse_Serialized_Name + "," +
                            DefaultHouse_Serialized_HouseFileName + "," +
                            DefaultHouse_Serialized_LocationsWithoutHidingPlaces + "," +
                            DefaultHouse_Serialized_LocationsWithHidingPlaces +
                        "}")
-                   .SetName("Test_House_CreateHouse_AndCheckErrorMessage_ForJsonException_WhenFileFormatIsInvalid - no PlayerStartingPoint");
+                   .SetName("Test_SavedGame_Deserialize_AndCheckErrorMessage_ForJsonException_WhenHouseFileFormatIsInvalid - no PlayerStartingPoint");
 
                 // File missing LocationsWithoutHidingPlaces
-                yield return new TestCaseData("JSON deserialization for type 'HideAndSeek.House' was missing required properties, including the following: LocationsWithoutHidingPlaces",
+                yield return new TestCaseData("JSON deserialization for type 'HideAndSeek.House' was missing required properties, " +
+                                              "including the following: LocationsWithoutHidingPlaces",
                         "{" +
                            DefaultHouse_Serialized_Name + "," +
                            DefaultHouse_Serialized_HouseFileName + "," +
                            DefaultHouse_Serialized_PlayerStartingPoint + "," +
                            DefaultHouse_Serialized_LocationsWithHidingPlaces +
                        "}")
-                   .SetName("Test_House_CreateHouse_AndCheckErrorMessage_ForJsonException_WhenFileFormatIsInvalid - no LocationsWithoutHidingPlaces");
+                   .SetName("Test_SavedGame_Deserialize_AndCheckErrorMessage_ForJsonException_WhenHouseFileFormatIsInvalid - no LocationsWithoutHidingPlaces");
 
                 // File missing LocationsWithHidingPlaces
-                yield return new TestCaseData("JSON deserialization for type 'HideAndSeek.House' was missing required properties, including the following: LocationsWithHidingPlaces",
+                yield return new TestCaseData("JSON deserialization for type 'HideAndSeek.House' was missing required properties, " +
+                                              "including the following: LocationsWithHidingPlaces",
                         "{" +
                            DefaultHouse_Serialized_Name + "," +
                            DefaultHouse_Serialized_HouseFileName + "," +
                            DefaultHouse_Serialized_PlayerStartingPoint + "," +
                            DefaultHouse_Serialized_LocationsWithoutHidingPlaces +
                        "}")
-                   .SetName("Test_House_CreateHouse_AndCheckErrorMessage_ForJsonException_WhenFileFormatIsInvalid - no LocationsWithHidingPlaces");
+                   .SetName("Test_SavedGame_Deserialize_AndCheckErrorMessage_ForJsonException_WhenHouseFileFormatIsInvalid - no LocationsWithHidingPlaces");
             }
         }
 
-        public static IEnumerable TestCases_For_Test_House_CreateHouse_AndCheckErrorMessage_ForInvalidDataException_WhenFileDataHasWhitespaceValue
+        public static IEnumerable TestCases_For_Test_SavedGame_Deserialize_AndCheckErrorMessage_ForInvalidDataException_WhenHouseFileDataInvalidValue
         {
             get
             {
-                // Invalid Name (whitespace)
-                yield return new TestCaseData("Cannot perform action because house name \" \" is invalid (is empty or contains only whitespace)",
+                // INVALID VALUE OF WHITESPACE
+                // Invalid Name - whitespace
+                yield return new TestCaseData("Cannot perform action because house name \" \" is invalid " +
+                                              "(is empty or contains only whitespace)",
                         "{" +
                            "\"Name\":\" \"" + "," +
                            DefaultHouse_Serialized_HouseFileName + "," +
@@ -545,10 +758,11 @@ namespace HideAndSeek
                            DefaultHouse_Serialized_LocationsWithoutHidingPlaces + "," +
                            DefaultHouse_Serialized_LocationsWithHidingPlaces +
                        "}")
-                   .SetName("Test_House_CreateHouse_AndCheckErrorMessage_ForInvalidDataException_WhenFileDataHasWhitespaceValue - invalid Name - whitespace");
+                   .SetName("Test_SavedGame_Deserialize_AndCheckErrorMessage_ForInvalidDataException_WhenHouseFileDataInvalidValue - invalid Name - whitespace");
 
-                // Invalid HouseFileName (whitespace)
-                yield return new TestCaseData("Cannot perform action because house file name \" \" is invalid (is empty or contains illegal characters, e.g. \\, /, or whitespace)",
+                // Invalid HouseFileName - whitespace
+                yield return new TestCaseData("Cannot perform action because house file name \" \" is invalid " +
+                                              "(is empty or contains illegal characters, e.g. \\, /, or whitespace)",
                         "{" +
                            DefaultHouse_Serialized_Name + "," +
                            "\"HouseFileName\":\" \"" + "," +
@@ -556,10 +770,11 @@ namespace HideAndSeek
                            DefaultHouse_Serialized_LocationsWithoutHidingPlaces + "," +
                            DefaultHouse_Serialized_LocationsWithHidingPlaces +
                        "}")
-                   .SetName("Test_House_CreateHouse_AndCheckErrorMessage_ForInvalidDataException_WhenFileDataHasWhitespaceValue - invalid HouseFileName - whitespace");
+                   .SetName("Test_SavedGame_Deserialize_AndCheckErrorMessage_ForInvalidDataException_WhenHouseFileDataInvalidValue - invalid HouseFileName - whitespace");
 
-                // Invalid PlayerStartingPoint (whitespace)
-                yield return new TestCaseData("Cannot perform action because player starting point location name \" \" is invalid (is empty or contains only whitespace)",
+                // Invalid PlayerStartingPoint - whitespace
+                yield return new TestCaseData("Cannot perform action because player starting point location name \" \" is invalid " +
+                                              "(is empty or contains only whitespace)",
                         "{" +
                            DefaultHouse_Serialized_Name + "," +
                            DefaultHouse_Serialized_HouseFileName + "," +
@@ -567,10 +782,11 @@ namespace HideAndSeek
                            DefaultHouse_Serialized_LocationsWithoutHidingPlaces + "," +
                            DefaultHouse_Serialized_LocationsWithHidingPlaces +
                        "}")
-                   .SetName("Test_House_CreateHouse_AndCheckErrorMessage_ForInvalidDataException_WhenFileDataHasWhitespaceValue - invalid PlayerStartingPoint - whitespace");
+                   .SetName("Test_SavedGame_Deserialize_AndCheckErrorMessage_ForInvalidDataException_WhenHouseFileDataInvalidValue - invalid PlayerStartingPoint - whitespace");
 
                 // Invalid LocationsWithoutHidingPlaces - Location name is invalid (whitespace)
-                yield return new TestCaseData("Cannot perform action because location name \" \" is invalid (is empty or contains only whitespace)",
+                yield return new TestCaseData("Cannot perform action because location name \" \" is invalid " +
+                                              "(is empty or contains only whitespace)",
                         "{" +
                            DefaultHouse_Serialized_Name + "," +
                            DefaultHouse_Serialized_HouseFileName + "," +
@@ -591,10 +807,11 @@ namespace HideAndSeek
                            "]" + "," +
                            DefaultHouse_Serialized_LocationsWithHidingPlaces +
                        "}")
-                   .SetName("Test_House_CreateHouse_AndCheckErrorMessage_ForInvalidDataException_WhenFileDataHasWhitespaceValue - LocationsWithoutHidingPlaces - invalid Location Name - whitespace");
+                   .SetName("Test_SavedGame_Deserialize_AndCheckErrorMessage_ForInvalidDataException_WhenHouseFileDataInvalidValue - LocationsWithoutHidingPlaces - invalid Location Name - whitespace");
 
                 // Invalid LocationsWithHidingPlaces - LocationWithHidingPlace name in invalid (whitespace)
-                yield return new TestCaseData("Cannot perform action because location name \" \" is invalid (is empty or contains only whitespace)",
+                yield return new TestCaseData("Cannot perform action because location name \" \" is invalid " +
+                                              "(is empty or contains only whitespace)",
                         "{" +
                            DefaultHouse_Serialized_Name + "," +
                            DefaultHouse_Serialized_HouseFileName + "," +
@@ -612,10 +829,11 @@ namespace HideAndSeek
                                 "}" +
                             "]" +
                        "}")
-                   .SetName("Test_House_CreateHouse_AndCheckErrorMessage_ForInvalidDataException_WhenFileDataHasWhitespaceValue - LocationsWithHidingPlaces - invalid Location Name - whitespace");
+                   .SetName("Test_SavedGame_Deserialize_AndCheckErrorMessage_ForInvalidDataException_WhenHouseFileDataInvalidValue - LocationsWithHidingPlaces - invalid Location Name - whitespace");
 
                 // Invalid LocationsWithHidingPlaces - LocationWithHidingPlace description is invalid (whitespace)
-                yield return new TestCaseData("Cannot perform action because hiding place \" \" is invalid (is empty or contains only whitespace)",
+                yield return new TestCaseData("Cannot perform action because hiding place \" \" is invalid " +
+                                              "(is empty or contains only whitespace)",
                         "{" +
                            DefaultHouse_Serialized_Name + "," +
                            DefaultHouse_Serialized_HouseFileName + "," +
@@ -633,89 +851,12 @@ namespace HideAndSeek
                                 "}" +
                             "]" +
                        "}")
-                   .SetName("Test_House_CreateHouse_AndCheckErrorMessage_ForInvalidDataException_WhenFileDataHasWhitespaceValue - LocationsWithHidingPlaces - invalid Location HidingPlace - whitespace");
-            }
-        }
+                   .SetName("Test_SavedGame_Deserialize_AndCheckErrorMessage_ForInvalidDataException_WhenHouseFileDataInvalidValue - LocationsWithHidingPlaces - invalid Location HidingPlace - whitespace");
 
-        public static IEnumerable TestCases_For_Test_House_CreateHouse_AndCheckErrorMessage_ForJsonException_WhenFileDataHasInvalidDirection
-        {
-            get
-            {
-                // Invalid LocationsWithoutHidingPlaces - exit Direction is invalid
-                yield return new TestCaseData("{" +
-                           DefaultHouse_Serialized_Name + "," +
-                           DefaultHouse_Serialized_HouseFileName + "," +
-                           DefaultHouse_Serialized_PlayerStartingPoint + "," +
-                           "\"LocationsWithoutHidingPlaces\":" +
-                           "[" +
-                                "{" +
-                                    "\"Name\":\"Hallway\"," +
-                                    "\"ExitsForSerialization\":" +
-                                    "{" +
-                                        "\"West\":\"Entry\"," +
-                                        "\"Up\":\"Attic\"" +
-                                    "}" +
-                                "}," +
-                                "{" +
-                                    "\"Name\":\"Entry\"," +
-                                    "\"ExitsForSerialization\":" +
-                                    "{" +
-                                        "\"Insideout\":\"Hallway\"" +
-                                    "}" +
-                                "}" +
-                            "]" + "," +
-                           "\"LocationsWithHidingPlaces\":" +
-                           "[" + // Must have at least one location with hiding place
-                                "{" +
-                                    "\"HidingPlace\":\"in a trunk\"," +
-                                    "\"Name\":\"Attic\"," +
-                                    "\"ExitsForSerialization\":" +
-                                    "{" +
-                                        "\"Down\":\"Hallway\"" +
-                                    "}" +
-                                "}" +
-                           "]" +
-                       "}")
-                   .SetName("Test_House_CreateHouse_AndCheckErrorMessage_ForJsonException_WhenFileDataHasInvalidDirection - LocationsWithoutHidingPlaces");
-
-                // Invalid LocationsWithHidingPlaces - exit direction is invalid
-                yield return new TestCaseData("{" +
-                            DefaultHouse_Serialized_Name + "," +
-                            DefaultHouse_Serialized_HouseFileName + "," +
-                            "\"PlayerStartingPoint\":\"Master Bedroom\"" + "," +
-                            "\"LocationsWithoutHidingPlaces\":" +
-                            "[" +
-                            "]" + "," +
-                            "\"LocationsWithHidingPlaces\":" +
-                            "[" +
-                                "{" +
-                                    "\"HidingPlace\":\"under the bed\"," +
-                                    "\"Name\":\"Master Bedroom\"," +
-                                    "\"ExitsForSerialization\":" +
-                                    "{" +
-                                        "\"East\":\"Master Bath\"" +
-                                    "}" +
-                                "}," +
-                                "{" +
-                                    "\"HidingPlace\":\"in the tub\"," +
-                                    "\"Name\":\"Master Bath\"," +
-                                    "\"ExitsForSerialization\":" +
-                                    "{" +
-                                        "\"Insideout\":\"Master Bedroom\"" +
-                                    "}" +
-                                "}" +
-                            "]" +
-                        "}")
-                   .SetName("Test_House_CreateHouse_AndCheckErrorMessage_ForJsonException_WhenFileDataHasInvalidDirection - LocationsWithHidingPlaces");
-            }
-        }
-
-        public static IEnumerable TestCases_For_Test_House_CreateHouse_AndCheckErrorMessage_ForInvalidDataException_WhenFileDataHasInvalidValue
-        {
-            get
-            {
+                // OTHER INVALID VALUES
                 // Invalid PlayerStartingPoint - not a Location
-                yield return new TestCaseData("Cannot perform action because player starting point location \"Dungeon\" is not a location in the house",
+                yield return new TestCaseData("Cannot perform action because player starting point location \"Dungeon\" " +
+                                              "is not a location in the house",
                         "{" +
                            DefaultHouse_Serialized_Name + "," +
                            DefaultHouse_Serialized_HouseFileName + "," +
@@ -723,7 +864,7 @@ namespace HideAndSeek
                            DefaultHouse_Serialized_LocationsWithoutHidingPlaces + "," +
                            DefaultHouse_Serialized_LocationsWithHidingPlaces +
                        "}")
-                   .SetName("Test_House_CreateHouse_AndCheckErrorMessage_ForInvalidDataException_WhenFileDataHasInvalidValue - invalid PlayerStartingPoint - nonexistent");
+                   .SetName("Test_SavedGame_Deserialize_AndCheckErrorMessage_ForInvalidDataException_WhenHouseFileDataInvalidValue - invalid PlayerStartingPoint - nonexistent");
 
                 // Invalid LocationsWithHidingPlaces - empty list
                 yield return new TestCaseData("Cannot perform action because locations with hiding places list is empty",
@@ -752,7 +893,7 @@ namespace HideAndSeek
                            "[" +
                            "]" +
                        "}")
-                   .SetName("Test_House_CreateHouse_AndCheckErrorMessage_ForInvalidDataException_WhenFileDataHasInvalidValue - invalid LocationsWithHidingPlaces - empty");
+                   .SetName("Test_SavedGame_Deserialize_AndCheckErrorMessage_ForInvalidDataException_WhenHouseFileDataInvalidValue - invalid LocationsWithHidingPlaces - empty");
 
                 // Invalid LocationsWithoutHidingPlaces - Location has no exits
                 yield return new TestCaseData("Cannot perform action because location \"Hallway\" must be assigned at least one exit",
@@ -792,14 +933,14 @@ namespace HideAndSeek
                            "]" + "," +
                            DefaultHouse_Serialized_LocationsWithHidingPlaces +
                        "}")
-                   .SetName("Test_House_CreateHouse_AndCheckErrorMessage_ForInvalidDataException_WhenFileDataHasInvalidValue - invalid LocationsWithoutHidingPlaces - no exits");
+                   .SetName("Test_SavedGame_Deserialize_AndCheckErrorMessage_ForInvalidDataException_WhenHouseFileDataInvalidValue - invalid LocationsWithoutHidingPlaces - no exits");
 
                 // Invalid LocationsWithHidingPlaces - LocationWithHidingPlace has no exits
                 yield return new TestCaseData("Cannot perform action because location \"Attic\" must be assigned at least one exit",
                         "{" +
                            DefaultHouse_Serialized_Name + "," +
                            DefaultHouse_Serialized_HouseFileName + "," +
-                           DefaultHouse_Serialized_PlayerStartingPoint + "," + 
+                           DefaultHouse_Serialized_PlayerStartingPoint + "," +
                            DefaultHouse_Serialized_LocationsWithoutHidingPlaces + "," +
                            "\"LocationsWithHidingPlaces\":" +
                            "[" +
@@ -812,10 +953,11 @@ namespace HideAndSeek
                                 "}" +
                            "]" +
                        "}")
-                   .SetName("Test_House_CreateHouse_AndCheckErrorMessage_ForInvalidDataException_WhenFileDataHasInvalidValue - invalid LocationsWithHidingPlaces - no exits");
+                   .SetName("Test_SavedGame_Deserialize_AndCheckErrorMessage_ForInvalidDataException_WhenHouseFileDataInvalidValue - invalid LocationsWithHidingPlaces - no exits");
 
                 // Invalid LocationsWithoutHidingPlaces - Location has nonexistent exit
-                yield return new TestCaseData("Cannot perform action because \"Hallway\" exit location \"Dungeon\" in direction \"Down\" does not exist",
+                yield return new TestCaseData("Cannot perform action because \"Hallway\" exit location \"Dungeon\" " +
+                                              "in direction \"Down\" does not exist",
                         "{" +
                            DefaultHouse_Serialized_Name + "," +
                            DefaultHouse_Serialized_HouseFileName + "," +
@@ -858,10 +1000,11 @@ namespace HideAndSeek
                            "]" + "," +
                            DefaultHouse_Serialized_LocationsWithHidingPlaces +
                        "}")
-                   .SetName("Test_House_CreateHouse_AndCheckErrorMessage_ForInvalidDataException_WhenFileDataHasInvalidValue - invalid LocationsWithoutHidingPlaces - nonexistent exit");
+                   .SetName("Test_SavedGame_Deserialize_AndCheckErrorMessage_ForInvalidDataException_WhenHouseFileDataInvalidValue - invalid LocationsWithoutHidingPlaces - nonexistent exit");
 
                 // Invalid LocationsWithHidingPlaces - LocationWithHidingPlace has nonexistent exit
-                yield return new TestCaseData("Cannot perform action because \"Bathroom\" exit location \"Dungeon\" in direction \"Down\" does not exist",
+                yield return new TestCaseData("Cannot perform action because \"Bathroom\" exit location \"Dungeon\" " +
+                                              "in direction \"Down\" does not exist",
                         "{" +
                            DefaultHouse_Serialized_Name + "," +
                            DefaultHouse_Serialized_HouseFileName + "," +
@@ -960,7 +1103,80 @@ namespace HideAndSeek
                                 "}" +
                            "]" +
                        "}")
-                   .SetName("Test_House_CreateHouse_AndCheckErrorMessage_ForInvalidDataException_WhenFileDataHasInvalidValue - invalid LocationsWithHidingPlaces - nonexistent exit");
+                   .SetName("Test_SavedGame_Deserialize_AndCheckErrorMessage_ForInvalidDataException_WhenHouseFileDataInvalidValue - invalid LocationsWithHidingPlaces - nonexistent exit");
+            }
+        }
+
+        public static IEnumerable TestCases_For_Test_SavedGame_Deserialize_AndCheckErrorMessage_ForJsonException_WhenHouseFileDataHasInvalidDirection
+        {
+            get
+            {
+                // Invalid LocationsWithoutHidingPlaces - exit Direction is invalid
+                yield return new TestCaseData("{" +
+                           DefaultHouse_Serialized_Name + "," +
+                           DefaultHouse_Serialized_HouseFileName + "," +
+                           DefaultHouse_Serialized_PlayerStartingPoint + "," +
+                           "\"LocationsWithoutHidingPlaces\":" +
+                           "[" +
+                                "{" +
+                                    "\"Name\":\"Hallway\"," +
+                                    "\"ExitsForSerialization\":" +
+                                    "{" +
+                                        "\"West\":\"Entry\"," +
+                                        "\"Up\":\"Attic\"" +
+                                    "}" +
+                                "}," +
+                                "{" +
+                                    "\"Name\":\"Entry\"," +
+                                    "\"ExitsForSerialization\":" +
+                                    "{" +
+                                        "\"Insideout\":\"Hallway\"" +
+                                    "}" +
+                                "}" +
+                            "]" + "," +
+                           "\"LocationsWithHidingPlaces\":" +
+                           "[" + // Must have at least one location with hiding place
+                                "{" +
+                                    "\"HidingPlace\":\"in a trunk\"," +
+                                    "\"Name\":\"Attic\"," +
+                                    "\"ExitsForSerialization\":" +
+                                    "{" +
+                                        "\"Down\":\"Hallway\"" +
+                                    "}" +
+                                "}" +
+                           "]" +
+                       "}")
+                   .SetName("Test_SavedGame_Deserialize_AndCheckErrorMessage_ForJsonException_WhenHouseFileDataHasInvalidDirection - LocationsWithoutHidingPlaces");
+
+                // Invalid LocationsWithHidingPlaces - exit direction is invalid
+                yield return new TestCaseData("{" +
+                            DefaultHouse_Serialized_Name + "," +
+                            DefaultHouse_Serialized_HouseFileName + "," +
+                            "\"PlayerStartingPoint\":\"Master Bedroom\"" + "," +
+                            "\"LocationsWithoutHidingPlaces\":" +
+                            "[" +
+                            "]" + "," +
+                            "\"LocationsWithHidingPlaces\":" +
+                            "[" +
+                                "{" +
+                                    "\"HidingPlace\":\"under the bed\"," +
+                                    "\"Name\":\"Master Bedroom\"," +
+                                    "\"ExitsForSerialization\":" +
+                                    "{" +
+                                        "\"East\":\"Master Bath\"" +
+                                    "}" +
+                                "}," +
+                                "{" +
+                                    "\"HidingPlace\":\"in the tub\"," +
+                                    "\"Name\":\"Master Bath\"," +
+                                    "\"ExitsForSerialization\":" +
+                                    "{" +
+                                        "\"Insideout\":\"Master Bedroom\"" +
+                                    "}" +
+                                "}" +
+                            "]" +
+                        "}")
+                   .SetName("Test_SavedGame_Deserialize_AndCheckErrorMessage_ForJsonException_WhenHouseFileDataHasInvalidDirection - LocationsWithHidingPlaces");
             }
         }
     }
